+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# [2.1.1-beta.2]
+
+* Activating a fundraiser account 
+* Reveal an account
+
 # [2.1.1-beta]
 
 * Fix for revelation of an account Ref #12